Still Trying To Fix Message Sending.

+++ C++/Libaries/Decrypt.h
@@ -1,9 +1,8 @@
 #pragma once
-#include <iostream>
 
 int funcGetIndexDecrypt(char letter, char def_alph[])
 {
-	int index = 0;
+	unsigned short index = 0;
 	while (letter != def_alph[index])
 	{
 		if (index < 52)
@@ -23,13 +22,13 @@ std::string funcDecrypt(int seed[], std::string text)
 	char def_alph[52]{ 'A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P','Q','R','S','T','U','V','W','X','Y','Z','a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z' };
 	char new_alph[52]{ 'A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P','Q','R','S','T','U','V','W','X','Y','Z','a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z' };
 	std::string enc_text = "";
-	for (int i = 0; i < 52; i++)
+	for (unsigned short i = 0; i < 52; i++)
 	{
 		char buffer = new_alph[i];
 		new_alph[i] = new_alph[seed[i]];
 		new_alph[seed[i]] = buffer;
 	}
-	for (int i = 0; i < text.size(); i++)
+	for (unsigned short i = 0; i < text.size(); i++)
 	{
 		if (std::isalpha(text[i]))
 		{
